+++ wakamiti-api/CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [2.10.0] - 2026-09-09
 
 ### Added
 - Support setup and tear-down operations at plan, feature and scenario levels.